Do some refactoring and tweak a couple of queries
