+++ services/QuillLMS/app/queries/snapshots/active_students_query.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Snapshots
-  class ActiveStudentsQuery < CountQuery
+  class ActiveStudentsQuery < ActivitySessionCountQuery
     def select_clause
       "SELECT COUNT(DISTINCT activity_sessions.user_id) AS count"
     end
