+++ services/QuillLMS/app/queries/snapshots/activities_completed_query.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Snapshots
-  class ActivitiesCompletedQuery < CountQuery
+  class ActivitiesCompletedQuery < ActivitySessionCountQuery
     def select_clause
       "SELECT COUNT(DISTINCT activity_sessions.id) AS count"
     end
