+++ services/QuillLMS/app/queries/snapshots/student_learning_hours_query.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Snapshots
-  class StudentLearningHoursQuery < CountQuery
+  class StudentLearningHoursQuery < ActivitySessionCountQuery
     def select_clause
       # timespent stores seconds
       "SELECT SUM(activity_sessions.timespent) / 3600.0 AS count"
