+++ services/QuillLMS/spec/queries/snapshots/active_classrooms_query_spec.rb
@@ -5,7 +5,7 @@ require 'rails_helper'
 module Snapshots
   describe ActiveClassroomsQuery do
     context 'external_api', :big_query_snapshot do
-      include_context 'Snapshots Count CTE'
+      include_context 'Snapshots Activity Session Count CTE'
 
       it { expect(results).to eq(count: num_classrooms) }
     end
