+++ services/QuillLMS/spec/queries/snapshots/activities_completed_query_spec.rb
@@ -5,7 +5,7 @@ require 'rails_helper'
 module Snapshots
   describe ActivitiesCompletedQuery do
     context 'external_api', :big_query_snapshot do
-      include_context 'Snapshots Count CTE'
+      include_context 'Snapshots Activity Session Count CTE'
 
       let(:num_completed_activities) { activity_sessions.count }
