+++ services/QuillLMS/spec/queries/snapshots/sentences_written_query_spec.rb
@@ -5,7 +5,7 @@ require 'rails_helper'
 module Snapshots
   describe SentencesWrittenQuery do
     context 'external_api', :big_query_snapshot do
-      include_context 'Snapshots Count CTE'
+      include_context 'Snapshots Activity Session Count CTE'
 
       let(:concept_results) { activity_sessions.map { |activity_session| create(:concept_result, activity_session: activity_session) } }
       let(:cte_records) { count_query_cte_records << concept_results }
